requirements.txt
streamlit>=1.30
pandas
PyPDF2
python-docx
fpdf2
openai
docx
